Change response structure
Thanks to the refactoring just done, if we now need to modify the
structure of a response we can do so directly in the frontend, through
the response_editor function parameter passed to RedisManager.request.

This commit exploits this feature to change the structure of responses
on the /coinglass/open_interest path and align them to the requests of
the contractor.